moving permissions logic to a separate file
