+++ app/src/main/java/com/example/weatherapp/MainActivity.kt
@@ -1,54 +1,39 @@
 package com.example.weatherapp
 
-import android.Manifest
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.compose.setContent
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
 import com.example.weatherapp.core.presentation.navigation.RootNavigationGraph
 import com.example.weatherapp.ui.theme.WeatherAppTheme
+import com.example.weatherapp.utils.CheckPermissions
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
-import com.google.accompanist.permissions.rememberMultiplePermissionsState
 import dagger.hilt.android.AndroidEntryPoint
 
-@OptIn(ExperimentalPermissionsApi::class)
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             WeatherAppTheme {
-                val permissionsState = rememberMultiplePermissionsState(
-                    permissions = listOf(
-                        Manifest.permission.ACCESS_COARSE_LOCATION,
-                        Manifest.permission.ACCESS_FINE_LOCATION
-                    )
-                ) { permissions ->
-                    if (permissions.values.any { !it }) {
+                CheckPermissions(
+                    onPermissionDenied = {
                         finishAffinity()
-                    }
-                }
-                if (permissionsState.allPermissionsGranted) {
-                    Surface(
-                        modifier = Modifier.fillMaxSize(), color = MaterialTheme.colors.background
-                    ) {
-                        RootNavigationGraph(
-                            navController = rememberNavController(), modifier = Modifier
-                        )
-                    }
-                } else {
-                    LaunchedEffect(key1 = true) {
-                        permissionsState.launchMultiplePermissionRequest()
-                    }
-                }
+                    },
+                    onPermissionGranted = {
+                        Surface(
+                            modifier = Modifier.fillMaxSize(),
+                            color = MaterialTheme.colors.background
+                        ) {
+                            RootNavigationGraph(
+                                navController = rememberNavController(), modifier = Modifier
+                            )
+                        }
+                    })
             }
         }
     }
